services/slash-bot/requirements.txt
boto3
discord.py
